pyproject.toml
[tool.poetry]
name = "cal-opt"
version = "0.1.0"
description = "Calendar leave optimiser"
authors = ["Ariel Richtman <ariel@richtman.com.au>"]
license = "GPL3"
homepage = "https://github.com/arichtman/cal-opt"
repository = "https://github.com/arichtman/cal-opt"
documentation = "https://arichtman.github.io/cal-opt/"
packages = [
    { include = "cal_opt", from = "src" },
]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
]

[tool.poetry.urls]
Changelog = "https://github.com/arichtman/cal-opt/releases"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^0.19.2"
click = "^8.0.4"
icalendar = "^4.0.9"
ics = "^0.7"

[tool.poetry.dev-dependencies]
mkdocs = "^1.2.2"
black = "^21.9b0"
commitizen = "^2.18.1"
pre-commit = "^2.15.0"
pytest = "^7.0.0"
safety = "^1.10.3"
pyupgrade = "^2.29.1"
nox = "^2021.10.1"
jake = "^1.4.5"
bandit = "^1.7.1"
adr-tools-python = "^1.0.3"

[tool.poetry.scripts]
cal-opt = "cal_opt.__main__:cli"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
target-version = ['py310']
include = '\.pyi?$'
